+++ Changelog.md
@@ -12,6 +12,7 @@ Features:
  * Inline Assembly: ``for`` and ``switch`` statements.
  * Inline Assembly: function definitions and function calls.
  * Code Generator: Added the Whiskers template system.
+ * Remove obsolete Why3 output.
 
 Bugfixes:
  * Code generator: Use ``REVERT`` instead of ``INVALID`` for generated input validation routines.